+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT94XBKWMHRZNJVJ9ATT0NH5"
+        "01KT997DZEKB74HR2ARTSSAXKY"
     )
     r"""The unique identifier of the data source"""