+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-04T11:04:28.064Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-04T12:19:55.618Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")